Simplify backprop by simplifying cache usage
Instead the storing z and a (the weighted sum of inputs and the
activation) of each layer together in each cache item, store z from one
layer and a from the previous layer together in each cache item.

This makes backward propagation simpler because at each layer we need
the output of the previous layer to compute dw, i.e. the derivatives
with respect to the cost function.